最终初始化版本

+++ src/main/java/com/wjdu/res/service/thread/GetPictureThread.java
@@ -27,28 +27,32 @@ public class GetPictureThread implements Callable<ImageDTO> {
     @Override
     public ImageDTO call() throws Exception {
         ImageDTO imageDTO = new ImageDTO();
-        URL url;
-        InputStream in;
-        FileOutputStream file;
-        String fileName = "";
         try {
             // 获取图片名
-            fileName = this.remoteSrc.substring(this.remoteSrc.lastIndexOf("/")).replace("/", "");
-            // 初始化url对象
-            url = new URL(this.remoteSrc);
-            URLConnection conn = url.openConnection();
-            // 初始化in对象，也就是获得url字节流
-            conn.setRequestProperty("User-Agent", "Mozilla/4.0 (compatible; MSIE 5.0; Windows NT; DigExt)");
-            in = conn.getInputStream();
-            byte[] bs = new byte[1024];
-            int len;
-            OutputStream os = new FileOutputStream(this.filePath + "\\" + fileName);
-            // 开始读取
-            while ((len = in.read(bs)) != -1) {
-                os.write(bs, 0, len);
+            String fileName = this.remoteSrc.substring(this.remoteSrc.lastIndexOf("/")).replace("/", "");
+            if(fileName.contains(".jpg")||fileName.contains(".gif")||fileName.contains(".png")){
+                fileName = fileName;
+            }else{
+                fileName = fileName+".jpg";
             }
-            os.close();
-            in.close();
+            File file=new File(this.filePath + "\\" + fileName);
+            if(!file.exists()){
+                // 初始化url对象
+                URL url = new URL(this.remoteSrc);
+                URLConnection conn = url.openConnection();
+                // 初始化in对象，也就是获得url字节流
+                conn.setRequestProperty("User-Agent", "Mozilla/4.0 (compatible; MSIE 5.0; Windows NT; DigExt)");
+                InputStream in = conn.getInputStream();
+                byte[] bs = new byte[1024];
+                int len;
+                OutputStream os = new FileOutputStream(this.filePath + "\\" + fileName);
+                // 开始读取
+                while ((len = in.read(bs)) != -1) {
+                    os.write(bs, 0, len);
+                }
+                os.close();
+                in.close();
+            }
             imageDTO.setRemoteShortSrc(this.remoteShortSrc);
             imageDTO.setRemoteSrc(this.remoteSrc);
             imageDTO.setLocalScr(this.filePath + "\\" + fileName);
